# swagger/swagger.yml
openapi: 3.0.1
info:
  title: Vechicle Rent API
  description: Bike Rent API for Node Server
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Authentication server (Local)
  - url: http://localhost:3000/api/v1
    description: API server v1 (Local)
...
paths:
...
  /vehicles/:
    post:
      tags:
        - Vehicle
      summary: Create vehicle
      operationId: Createvehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreate"
      responses:
        "200":
          description: Successfull Response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IAPIResponse"
                  - properties:
                      body:
                        type: object
                        properties:
                          uid:
                            type: string
                          id:
                            type: integer
      security:
        - JWT: []
...
components:
  securitySchemes:
    JWT:
      type: apiKey
      name: Authorization
      in: header
...
  schemas:
...
    IAPIResponse:
      type: object
      properties:
        success:
          type: boolean
          readOnly: true
        message:
          type: string
          readOnly: true
        body:
          oneOf:
            - type: object
          readOnly: true
...
    VehicleCreate:
      type: object
      properties:
        sku:
          type: string
          format: uuid
        fuel_type_id:
          type: integer
        vehicle_no:
          type: string
        hourly_rent:
          type: integer
        daily_rent:
          type: integer
        vehicle_type:
          type: integer
        model:
          type: string
        make:
          type: string
        vehicle_variant_id:
          type: integer
...